web/testimage/views.py
from rest_framework import generics
from .models import Image
from .serializers import ImageSerializer
from rest_framework.permissions import IsAuthenticated
from knox.auth import TokenAuthentication
from rest_framework.authentication import SessionAuthentication, BasicAuthentication
from rest_framework.parsers import MultiPartParser, FormParser
from rest_framework.response import Response

# Create your views here.

class ImageListCreateView(generics.ListCreateAPIView):
    queryset = Image.objects.all()
    serializer_class = ImageSerializer
    
    
image_list_create_view = ImageListCreateView.as_view()